updated App.jsx and Router.jsx, moved Navigation folder to components
