+++ src/App.test.jsx
@@ -1,6 +1,5 @@
 import React from "react";
-import Enzyme, { shallow } from "enzyme";
-import EnzymeAdapter from "enzyme-adapter-react-16";
+import { shallow } from "enzyme";
 import App from "./App";
 import FAQ from "./pages/FAQ";
 import Information from "./pages/Information";
@@ -8,14 +7,6 @@ import Profile from "./pages/Profile";
 import Registration from "./pages/Registration";
 import AccordionList from "./components/AccordionList";
 import NavBar from "./components/NavBar";
-
-// Set up enzyme's react adapter
-Enzyme.configure({ adapter: new EnzymeAdapter() });
-
-// Factory function to create a ShallowWrapper for the App Component.
-//   @function setup
-//   @returns {ShallowWrapper}
-// const setup = () => shallow(<App />);
 
 const findByTestAttr = (wrapper, val) => wrapper.find(`[data-test="${val}"]`);
