+++ src/router/Router.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import NavBar from "../components/NavBar";
+import Navigation from "../components/Navigation";
 import Information from "../pages/Information";
 import Profile from "../pages/Profile";
 import FAQ from "../pages/FAQ";
@@ -8,9 +8,7 @@ import Registration from "../pages/Registration";
 
 const Router = () => (
   <BrowserRouter basename="/HacktoberFest">
-    <div className="App NavBar">
-      <NavBar />
-    </div>
+    <Navigation />
     <Switch>
       <Route path="/" exact component={Information} />
       <Route path="/registration" component={Registration} />
